refactor(git): WorkingDirectory 를 도메인 모듈로 분할
952줄 단일 파일을 7개 파일로 분할했다 — WorkingDirectory 530 ·
FileItem 130 · HunkStagingPanel 78 · StagedDiffReviewPanel 78 ·
SensitiveFilesDialog 75 · constants 46 · types 37.
JSX·props·훅 호출 순서를 한 글자도 바꾸지 않았다(rules-of-hooks 보존).
정의는 손으로 옮기지 않고 sed 라인 슬라이스로 이동했다.

레이아웃이 Task 1(notifications)과 다르다. components/notifications/ 는
컴포넌트 1개짜리 디렉토리라 types.ts·constants.ts 를 루트에 둬도 모호하지
않았지만, components/git/ 은 12개 컴포넌트가 공유하는 디렉토리다. 같은
상황의 선례가 이미 있다 — components/usage/llm-access/ 는 여러 컴포넌트가
있는 usage/ 안에서 LLMAccessSettings 하나를 쪼개면서 메인은 부모 루트에
두고 부품만 kebab-case 중첩 디렉토리로 내렸다. 그 배치를 따른다.

메인 파일이 제자리에 남으므로 __tests__/WorkingDirectory.test.tsx:3 의
'../WorkingDirectory' 와 index.ts 의 './WorkingDirectory' 가 글자 그대로
유효하다. 패키지 승격(X.tsx → X/index.tsx)은 두 경로를 디렉토리 해석에
의존하게 만들 뿐 이득이 없어 하지 않았다 — Task 1 에서도 생략한 단계다.

서브컴포넌트 4개는 index.ts 가 재노출하지 않는다. 소비자가 메인 하나뿐이고
(실측) claude-sessions/TranscriptViewer 선례를 따른다. 원본에서 전부 지역
함수였으므로 export 를 추가했다.

tsc 의 noUnusedLocals 가 이번에는 지적 0건이다. Task 1 에서 3건(Webhook·
Settings·testChannel)을 잡았던 것과 대비되는데, 심볼 카운트를 근사치로만
쓰고 import 목록을 소비 지점에서 역산한 결과다. 설정이 실제로 켜져 있는지
(tsconfig.json:15) 확인한 뒤 통과로 판정했다.

게이트: tsc 0 · ESLint 0(--max-warnings=0) · vitest 205 파일 4,365 테스트
전부 통과 · build 성공.
