+++ src/dashboard/src/components/git/WorkingDirectory.tsx
@@ -1,9 +1,6 @@
-import { useState, useMemo } from 'react'
+import { useState } from 'react'
 import {
   FileEdit,
-  FilePlus,
-  FileX,
-  FileQuestion,
   Check,
   Plus,
   Minus,
@@ -14,437 +11,18 @@ import {
   Sparkles,
   Loader2,
   ArrowUp,
-  Eye,
   ChevronDown,
   ChevronRight,
-  ShieldAlert,
-  Shield,
-  AlertTriangle,
 } from 'lucide-react'
 import { cn } from '../../lib/utils'
-import { GitWorkingStatus, GitStatusFile, FileStatusType, DraftCommit, DiffHunk } from '../../stores/git'
 import { groupFilesByPattern, draftCommitsToFileGroups, FileGroup } from '../../utils/gitGrouping'
 import { FileGroupCard } from './FileGroup'
-import { checkSensitiveFile, filterSensitiveFiles, type SensitivityLevel } from '../../utils/gitSafetyPatterns'
-import { calcDiffStats } from '../../utils/diffParser'
-
-type ViewMode = 'list' | 'grouped'
-
-interface WorkingDirectoryProps {
-  workingStatus: GitWorkingStatus | null
-  isLoading: boolean
-  onRefresh: () => void
-  onStageFiles: (paths: string[]) => Promise<boolean>
-  onStageAll: () => Promise<boolean>
-  onUnstageFiles: (paths: string[]) => Promise<boolean>
-  onUnstageAll: () => Promise<boolean>
-  onCommit: (message: string) => Promise<boolean>
-  onCommitAndPush?: (message: string) => Promise<boolean>
-  // Diff
-  onFetchFileDiff: (filePath: string, staged: boolean) => Promise<string>
-  fileDiffs: Record<string, string>
-  isLoadingDiff: boolean
-  // Staged diff review
-  onFetchStagedDiff: () => Promise<string | null>
-  stagedDiff: string | null
-  // Hunk staging
-  onFetchFileHunks: (filePath: string, staged: boolean) => Promise<DiffHunk[]>
-  onStageHunks: (filePath: string, hunkIndices: number[]) => Promise<boolean>
-  fileHunks: Record<string, DiffHunk[]>
-  // LLM Draft Commits
-  draftCommits: DraftCommit[]
-  isGeneratingDrafts: boolean
-  onGenerateDrafts: () => Promise<DraftCommit[]>
-  onClearDrafts: () => void
-}
-
-const statusIcons: Record<FileStatusType, typeof FileEdit> = {
-  modified: FileEdit,
-  added: FilePlus,
-  deleted: FileX,
-  renamed: FileEdit,
-  untracked: FileQuestion,
-  staged: Check,
-}
-
-const statusColors: Record<FileStatusType, string> = {
-  modified: 'text-yellow-500',
-  added: 'text-green-500',
-  deleted: 'text-red-500',
-  renamed: 'text-blue-500',
-  untracked: 'text-gray-400',
-  staged: 'text-green-500',
-}
-
-const statusLabels: Record<FileStatusType, string> = {
-  modified: 'M',
-  added: 'A',
-  deleted: 'D',
-  renamed: 'R',
-  untracked: 'U',
-  staged: 'S',
-}
-
-const sensitivityIcons: Record<SensitivityLevel, typeof Shield | null> = {
-  danger: ShieldAlert,
-  warning: AlertTriangle,
-  safe: null,
-}
-
-const sensitivityColors: Record<SensitivityLevel, string> = {
-  danger: 'text-red-500',
-  warning: 'text-yellow-500',
-  safe: '',
-}
-
-// =============================================================================
-// FileItem with diff preview, unstage, and safety badge
-// =============================================================================
-
-function FileItem({
-  file,
-  selected,
-  onSelect,
-  onStage,
-  onUnstage,
-  onToggleDiff,
-  showDiff,
-  diffContent,
-  isLoadingDiff,
-}: {
-  file: GitStatusFile
-  selected: boolean
-  onSelect: (path: string) => void
-  onStage?: (path: string) => void
-  onUnstage?: (path: string) => void
-  onToggleDiff: (path: string) => void
-  showDiff: boolean
-  diffContent?: string
-  isLoadingDiff: boolean
-}) {
-  const Icon = statusIcons[file.status]
-  const sensitivity = checkSensitiveFile(file.path)
-  const SafetyIcon = sensitivityIcons[sensitivity.level]
-
-  return (
-    <div>
-      <div
-        className={cn(
-          'flex items-center gap-2 px-3 py-2 hover:bg-gray-50 dark:hover:bg-gray-700/50 rounded-lg cursor-pointer group',
-          selected && 'bg-primary-50 dark:bg-primary-900/20'
-        )}
-        onClick={() => onToggleDiff(file.path)}
-      >
-        <input
-          type="checkbox"
-          checked={selected}
-          onChange={() => onSelect(file.path)}
-          className="w-4 h-4 rounded border-gray-300 dark:border-gray-600 text-primary-600 focus:ring-primary-500"
-          onClick={(e) => e.stopPropagation()}
-        />
-        <Icon className={cn('w-4 h-4', statusColors[file.status])} />
-        <span className="flex-1 text-sm truncate text-gray-700 dark:text-gray-300">
-          {file.path}
-        </span>
-        {SafetyIcon && (
-          <span title={sensitivity.reason}>
-            <SafetyIcon className={cn('w-4 h-4', sensitivityColors[sensitivity.level])} />
-          </span>
-        )}
-        <span
-          className={cn(
-            'text-xs font-bold px-1.5 py-0.5 rounded',
-            statusColors[file.status]
-          )}
-        >
-          {statusLabels[file.status]}
-        </span>
-        {/* Stage button for unstaged files */}
-        {onStage && !file.staged && (
-          <button
-            onClick={(e) => {
-              e.stopPropagation()
-              onStage(file.path)
-            }}
-            className="opacity-0 group-hover:opacity-100 p-1 hover:bg-gray-200 dark:hover:bg-gray-600 rounded transition-opacity"
-            title="Stage this file"
-          >
-            <Plus className="w-4 h-4 text-green-500" />
-          </button>
-        )}
-        {/* Unstage button for staged files */}
-        {onUnstage && file.staged && (
-          <button
-            onClick={(e) => {
-              e.stopPropagation()
-              onUnstage(file.path)
-            }}
-            className="opacity-0 group-hover:opacity-100 p-1 hover:bg-gray-200 dark:hover:bg-gray-600 rounded transition-opacity"
-            title="Unstage this file"
-          >
-            <Minus className="w-4 h-4 text-red-500" />
-          </button>
-        )}
-      </div>
-      {/* Inline diff preview */}
-      {showDiff && (
-        <div className="ml-6 mr-3 mb-2 bg-gray-900 rounded-lg overflow-hidden text-xs max-h-60 overflow-y-auto">
-          {isLoadingDiff ? (
-            <div className="flex items-center gap-2 p-3 text-gray-400">
-              <Loader2 className="w-3 h-3 animate-spin" />
-              Loading diff...
-            </div>
-          ) : diffContent ? (
-            <div>
-              {diffContent.split('\n').map((line, i) => (
-                <div
-                  key={i}
-                  className={cn(
-                    'px-3 py-0.5',
-                    line.startsWith('+') && !line.startsWith('+++') && 'bg-green-900/30 text-green-400',
-                    line.startsWith('-') && !line.startsWith('---') && 'bg-red-900/30 text-red-400',
-                    line.startsWith('@@') && 'bg-blue-900/30 text-blue-400',
-                    !line.startsWith('+') && !line.startsWith('-') && !line.startsWith('@@') && 'text-gray-500'
-                  )}
-                >
-                  {line || ' '}
-                </div>
-              ))}
-            </div>
-          ) : (
-            <div className="p-3 text-gray-500">No diff available</div>
-          )}
-        </div>
-      )}
-    </div>
-  )
-}
-
-// =============================================================================
-// Staged Diff Review Panel
-// =============================================================================
-
-function StagedDiffReviewPanel({
-  stagedDiff,
-  isLoading,
-  onFetch,
-}: {
-  stagedDiff: string | null
-  isLoading: boolean
-  onFetch: () => void
-}) {
-  const [isExpanded, setIsExpanded] = useState(false)
-  const stats = useMemo(() => stagedDiff ? calcDiffStats(stagedDiff) : null, [stagedDiff])
-
-  return (
-    <div className="bg-white dark:bg-gray-800 rounded-xl border border-blue-200 dark:border-blue-800 overflow-hidden">
-      <button
-        onClick={() => {
-          if (!isExpanded && !stagedDiff) onFetch()
-          setIsExpanded(!isExpanded)
-        }}
-        className="w-full px-4 py-3 flex items-center justify-between bg-blue-50 dark:bg-blue-900/20 hover:bg-blue-100 dark:hover:bg-blue-900/30 transition-colors"
-      >
-        <div className="flex items-center gap-2">
-          <Eye className="w-4 h-4 text-blue-600 dark:text-blue-400" />
-          <span className="font-medium text-blue-700 dark:text-blue-400 text-sm">
-            Review Staged Changes
-          </span>
-          {stats && (
-            <span className="text-xs text-gray-500 dark:text-gray-400">
-              ({stats.files} file{stats.files !== 1 ? 's' : ''}, +{stats.additions} -{stats.deletions})
-            </span>
-          )}
-        </div>
-        {isExpanded ? (
-          <ChevronDown className="w-4 h-4 text-blue-600 dark:text-blue-400" />
-        ) : (
-          <ChevronRight className="w-4 h-4 text-blue-600 dark:text-blue-400" />
-        )}
-      </button>
-      {isExpanded && (
-        <div className="max-h-96 overflow-y-auto bg-gray-900 text-xs">
-          {isLoading ? (
-            <div className="flex items-center gap-2 p-4 text-gray-400">
-              <Loader2 className="w-4 h-4 animate-spin" />
-              Loading staged diff...
-            </div>
-          ) : stagedDiff ? (
-            stagedDiff.split('\n').map((line, i) => (
-              <div
-                key={i}
-                className={cn(
-                  'px-4 py-0.5',
-                  line.startsWith('+') && !line.startsWith('+++') && 'bg-green-900/30 text-green-400',
-                  line.startsWith('-') && !line.startsWith('---') && 'bg-red-900/30 text-red-400',
-                  line.startsWith('@@') && 'bg-blue-900/30 text-blue-400',
-                  line.startsWith('diff --git') && 'bg-gray-800 text-gray-300 font-bold mt-2',
-                  !line.startsWith('+') && !line.startsWith('-') && !line.startsWith('@@') && !line.startsWith('diff') && 'text-gray-500'
-                )}
-              >
-                {line || ' '}
-              </div>
-            ))
-          ) : (
-            <div className="p-4 text-gray-500">No staged changes</div>
-          )}
-        </div>
-      )}
-    </div>
-  )
-}
-
-// =============================================================================
-// Sensitive Files Confirmation Dialog
-// =============================================================================
-
-function SensitiveFilesDialog({
-  dangers,
-  warnings,
-  onConfirm,
-  onCancel,
-}: {
-  dangers: { path: string; reason: string }[]
-  warnings: { path: string; reason: string }[]
-  onConfirm: () => void
-  onCancel: () => void
-}) {
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50">
-      <div className="bg-white dark:bg-gray-800 rounded-xl shadow-xl w-full max-w-md p-6">
-        <div className="flex items-center gap-2 mb-4">
-          <ShieldAlert className="w-6 h-6 text-red-500" />
-          <h3 className="text-lg font-semibold text-gray-900 dark:text-white">
-            Sensitive Files Detected
-          </h3>
-        </div>
-
-        {dangers.length > 0 && (
-          <div className="mb-3">
-            <p className="text-sm font-medium text-red-600 dark:text-red-400 mb-1">Dangerous files:</p>
-            {dangers.map(({ path, reason }) => (
-              <div key={path} className="flex items-center gap-2 text-xs text-red-500 ml-2 py-0.5">
-                <ShieldAlert className="w-3 h-3" />
-                <span className="">{path}</span>
-                <span className="text-gray-400">- {reason}</span>
-              </div>
-            ))}
-          </div>
-        )}
-
-        {warnings.length > 0 && (
-          <div className="mb-3">
-            <p className="text-sm font-medium text-yellow-600 dark:text-yellow-400 mb-1">Warning files:</p>
-            {warnings.map(({ path, reason }) => (
-              <div key={path} className="flex items-center gap-2 text-xs text-yellow-500 ml-2 py-0.5">
-                <AlertTriangle className="w-3 h-3" />
-                <span className="">{path}</span>
-                <span className="text-gray-400">- {reason}</span>
-              </div>
-            ))}
-          </div>
-        )}
-
-        <p className="text-sm text-gray-500 dark:text-gray-400 mt-4 mb-4">
-          Are you sure you want to stage these files?
-        </p>
-
-        <div className="flex justify-end gap-3">
-          <button
-            onClick={onCancel}
-            className="px-4 py-2 text-sm font-medium text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 rounded-lg transition-colors"
-          >
-            Cancel
-          </button>
-          <button
-            onClick={onConfirm}
-            className="px-4 py-2 text-sm font-medium text-white bg-red-600 hover:bg-red-700 rounded-lg transition-colors"
-          >
-            Stage Anyway
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-// =============================================================================
-// Hunk Staging UI
-// =============================================================================
-
-function HunkStagingPanel({
-  hunks,
-  onStageHunks,
-  isLoading,
-}: {
-  hunks: DiffHunk[]
-  onStageHunks: (indices: number[]) => Promise<boolean>
-  isLoading: boolean
-}) {
-  const [selectedHunks, setSelectedHunks] = useState<Set<number>>(new Set())
-
-  const toggleHunk = (idx: number) => {
-    setSelectedHunks((prev) => {
-      const next = new Set(prev)
-      if (next.has(idx)) next.delete(idx)
-      else next.add(idx)
-      return next
-    })
-  }
-
-  const handleStage = async () => {
-    if (selectedHunks.size === 0) return
-    const success = await onStageHunks(Array.from(selectedHunks))
-    if (success) setSelectedHunks(new Set())
-  }
-
-  return (
-    <div className="ml-6 mr-3 mb-2 space-y-2">
-      {hunks.map((hunk) => (
-        <div key={hunk.index} className="bg-gray-900 rounded-lg overflow-hidden">
-          <div className="flex items-center gap-2 px-3 py-1.5 bg-gray-800">
-            <input
-              type="checkbox"
-              checked={selectedHunks.has(hunk.index)}
-              onChange={() => toggleHunk(hunk.index)}
-              className="w-3.5 h-3.5 rounded border-gray-600 text-blue-500 focus:ring-blue-500"
-            />
-            <span className="text-xs text-blue-400">{hunk.header}</span>
-          </div>
-          <div className="text-xs max-h-32 overflow-y-auto">
-            {hunk.content.split('\n').map((line, i) => (
-              <div
-                key={i}
-                className={cn(
-                  'px-3 py-0.5',
-                  line.startsWith('+') && 'bg-green-900/30 text-green-400',
-                  line.startsWith('-') && 'bg-red-900/30 text-red-400',
-                  line.startsWith(' ') && 'text-gray-500'
-                )}
-              >
-                {line || ' '}
-              </div>
-            ))}
-          </div>
-        </div>
-      ))}
-      {selectedHunks.size > 0 && (
-        <button
-          onClick={handleStage}
-          disabled={isLoading}
-          className="flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium text-white bg-blue-600 hover:bg-blue-700 rounded-lg transition-colors disabled:opacity-50"
-        >
-          <Plus className="w-3.5 h-3.5" />
-          Stage {selectedHunks.size} Hunk{selectedHunks.size !== 1 ? 's' : ''}
-        </button>
-      )}
-    </div>
-  )
-}
-
-// =============================================================================
-// Main WorkingDirectory Component
-// =============================================================================
+import { filterSensitiveFiles } from '../../utils/gitSafetyPatterns'
+import type { WorkingDirectoryProps, ViewMode } from './working-directory/types'
+import { FileItem } from './working-directory/FileItem'
+import { StagedDiffReviewPanel } from './working-directory/StagedDiffReviewPanel'
+import { SensitiveFilesDialog } from './working-directory/SensitiveFilesDialog'
+import { HunkStagingPanel } from './working-directory/HunkStagingPanel'
 
 export function WorkingDirectory({
   workingStatus,
